Significant browsing speed improvement
The details tree was rendering a full prettyprint with all new lines etc
for every entry.  This was, in my testing, huge for some objects
producing some 664k of text just for a single attribute.  Pretty printer
has been scaled back to give a more reasonable default that also
truncates the text after 300 chars for now.