Added an dedicated task executor
This handles tasks that are being executed and notified clients via
signals.  The main controller now uses this to determine the cursor type
to show.  Ths type filter is also now using the executor so that it
doesn't block the whole GUI as it finds out what types to show in the
drop-down list.